// desktop/src/types/mutations/node-reaction-create.ts
export type NodeReactionCreateMutationInput = {
  type: 'node_reaction_create';
  userId: string;
  nodeId: string;
  reaction: string;
};

export type NodeReactionCreateMutationOutput = {
  success: boolean;
};

declare module '@/types/mutations' {
  interface MutationMap {
    node_reaction_create: {
      input: NodeReactionCreateMutationInput;
      output: NodeReactionCreateMutationOutput;
    };
  }
}